server/requirements.txt

Django==5.0.6
gunicorn==23.0.0
sqlparse==0.5.3
asgiref==3.9.1
psycopg==3.2.9
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.7.0
django-storages
celery==5.3.4
redis==5.0.1
kombu==5.3.4
django-redis==5.4.0
django-celery-beat==2.6.0 
python-dotenv==1.1.1  
python-dateutil==2.9.0.post0
pillow==11.3.0
typing_extensions==4.14.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

boto3==1.34.25
google-api-python-client==2.108.0
botocore==1.34.25
jmespath==1.0.1
s3transfer==0.10.0
packaging==25.0  
six==1.17.0
supabase
twilio==8.11.0

PyJWT==2.8.0
firebase-admin==6.5.0









